instancetype: Require Guest within MemoryInstancetype
CPUInstancetype and MemoryInstancetype are both required attributes of
VirtualMachineInstancetypeSpec however only CPUInstancetype would
previously require that an amount of guest visible resources were
provided through its' Guest attribute.

This change corrects this for MemoryInstancetype ensuring that an amount
of guest visible memory is always provided with an instancetype as per
the original design.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>