e2e: improve the vm pool tests
The VM pool suite contains tests with running VMs. These
tests only observe the statuses of the VMIs and the VM pool
resources. They don't interact directly with the guest OS.

Therefore we can use a diskless approach when a VM runs
the "no bootable device" phase.

It will save the time it takes to setup the container-disk.

Also, the minimum amount of RAM required to run SeaBIOS is 1M,
I've decided to put 2Mi to be on the safe side.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>