libnet: avoid WithCloudInitNoCloudNetworkData's boolean arg
A casual reader finds it hard to understand what is "false" in
    WithCloudInitNoCloudNetworkData(networkData, false)
This can be solved by introducing a named argument or even an
enumeration for encoding. However, none of our tests wants to set the
"true" value, so let us just drop the b64Encoding argument and make the
function and its users simpler.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>