Update SMBios: assign KVM Virtual Machine to product
The current product value for SMBios is None, but this causes problem
when using `systemd-ditect-virt` command to differentiate the current
virtualization environment on arm virtual machines. As this command
reads /sys/class/dmi/id/product_name to get the real value.

Considering kubevirt is currently technically based on KVM (but not
xen yet[1]), let's fix this by assigning 'KVM Virtual Machine'
to 'product'.

[1] https://groups.google.com/g/kubevirt-dev/c/C6cUgzTOsVg

Signed-off-by: Fei Li <lifei.shirley@bytedance.com>