virt-controller, vmi: Move VMI object modification out of status update
The VMI controller reconciling execution has two steps which have been
separated by methods: sync and status-update.
The sync is suppose to mutate the VMI metadata and spec, while the
status-update is suppose to mutate the status section.

There is a 3rd step, where the VMI is being actually updated against the
API server. This step was part of the status-update, until this change.

Updating the VMI object on the API server is now placed at the same
level as the sync and status-update calls.
It should be now clearer what is done at each step.

Signed-off-by: Edward Haas <edwardh@redhat.com>