Stop including all [storage-req] tests in other lanes
Tests that require advanced/special storage capabilities are marked with `[storage-req]`;
Currently, [storage-req] tests from all SIGs are added to each lane that uses rook-ceph.
We might want to revisit that, as it does not make much sense for sig-migrations (for example)
to run sig-storage tests.
The sig-migrations tests which have [storage-req] will continue to run as these are a subset.

If and when we introduce other less "advanced" storage lanes
we could capitalize on [storage-req] to skip these.

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>