Revert "Merge pull request #7986 from iholder-redhat/fix/MigrationForVMX"
This reverts commit f963215842eb86dce271d1ae4bfc320e97079915, reversing
changes made to 9b9848d90d6da884e253e0280a2b68973562adac.

Reverting as this patch requires at least 1 node to have
the <counter name='tsc' /> capability to start a Windows VM
with Hyper-V enlightenments enabled, this is a regression
as before this patch this was required only on migration.

Signed-off-by: Antonio Cardace <acardace@redhat.com>