errcheck, network, netstat, test: always handle errors
Handling all errors in the `network/setup` package was trickier than the
rest of the network package because some functions ignored the returned
error on purpose.

To handle this error, it was needed to understand the cached files were
not actually being written to the FS because they were attempting to use
the *real* ownership manager - which would set `qemu` as the owner of
the cached files - this user does not exist in the unit test
environment.

Thus, it was needed to invoke `dutils.MockDefaultOwnershipManager()`, to
replace the ownership manager with a no-op.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>