e2e tests, libnode: only uncordon node once
`libnode` is relying on `kubectl uncordon` to make the node
schedulable again - this fails on the conformance lanes executed on the
kubevirtci project, since the sonobuoy container where the conformance
tests are executed do **not** have the kubectl binary.

Uncordoning a node is, quite literaly, making it schedulable - something
which is achieved by setting the `node.Spec.Unschedulable` flag to
false, and is already done in the stategic-merge patch command being sent to
the API some lines below, in [0].

Thus, we just need to make sure we also send the strategic merge PATCH
opeation for uncordoning a node.

[0] - https://github.com/kubevirt/kubevirt/blob/cdbc0a94a93ec5a758cde487edc62499ae2ff84f/tests/libnode/node.go#L107

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>