instancetype: Move ControllerRevision creation to VM mutation webhook
ControllerRevisions for instancetypes and preferences have until now
been captured by the VirtualMachine controller when creating an initial
VirtualMachineInstance at runtime. While this has worked well it leaves
a sustainable window where users could easily update or even remove the
original instancetype or preference leading to issues when later
launching the VirtualMachine.

This change seeks to remove this race window entirely by moving the
creation of these ControllerRevisions to the VirtualMachine mutation
webhook, early on in the admission process of the VirtualMachine. This
ensures a copy of any referenced instancetype or preference is captured
early and referenced from the VirtualMachine throughout its' lifecycle.

As the VirtualMachine mutation webhook constructs and applies a patch to
the VirtualMachine already some code that previously facilitated this
from the VirtualMachine controller can now be removed from the
instancetype module. This includes removing the passing of
ControllerRevisions as return values within the module.

For now an OwnerReference is no longer retained from the
ControllerRevision to the VirtualMachine as the VirtualMachine available
within the mutation webhook has not been fully created with a UID etc.
This will be corrected in a following commit by new logic within the
VirtualMachine controller.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>