Make virtctl ssh/pf work with dots in vm name
Currently, virtual machines may have dots in their VM name. This
currently breaks the logic for parsing portforward/ssh/scp targets, as
they assume that no dots can appear in the vm name.

To fix this, we assume that only the last part of the target name
(after splitting on dots) is the namespace. A namespace cannot contain
dots, so that seems like a safe assumption.

Signed-off-by: roblabla <unfiltered@roblab.la>