tests, sriov: Split the helper that creates two VMI/s
The helper function that creates two VMI/s for connectivity check is
removed in favor of one that handles each VMI separately.

The original helper had an odd signature which made it hard to
"trust" without looking at the implementation.

Once the separation has been done, the new helper looks much simpler and
can easily be created as a regular function.

There is however a cost to this change: The tests need to explicitly
create and wait for the two VMI/s, which adds caller code.
One can argue that it also makes the tests more explicit, exposing
exactly what is done without hiding it too much.
But all this is a matter of balance, there is probably no good answer
that will satisfy all.

Signed-off-by: Edward Haas <edwardh@redhat.com>