tests, sriov: Wait for VMI removal at test teardown
SR-IOV tests execution on CI has shown flakes where the VF configuration
has been modified after assigning the device to the pod.

The primary suspect is a CNI DELETE instruction from a previous pod
which is somehow racing with the re-assignment of the device to the new
pod.

In order to mitigate this suspected issue (and as side effect, prove it)
this change assures the VMI/s used with SR-IOV are deleted and absent
before starting the next test.

The issue is tracked on the sriov-cni project:
https://github.com/k8snetworkplumbingwg/sriov-cni/issues/219

Signed-off-by: Edward Haas <edwardh@redhat.com>