kubevirt/builder: Update sonobuoy version on dockerfile
This image is used for the periodic nightly job,
which runs kubevirt conformance.
Since v0.56.9 and the related changes are not backward compatible,
(new resultsDir and folder collecting feature)
we will need to update the Dockerfile, hack/dockerized once there is a new image
and then take the relevant changes of kubevirt and kubevirtci.

Signed-off-by: Or Shoval <oshoval@redhat.com>