e2e tests, migration: provide node drain + taint in a single func
The test framework performs teardown after each run test - part of
which, is to reset the taints and spec.Unschedulable attributes on the
nodes **if** they were previously tainted with the `NoExecute` taint.

Providing a function that when draining also applies the `NoExecute`
taint makes the developer life easier, since the test framework will
afterward reset the corresponding node.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>