Unlock and delete vmsnapshot only if content deleted if needed
Remove the vmsnapshotcontent finalizer in update content reconcile loop.
This is in order to prevent it being deleted before the source is
unfrozen.
prevent from vmsnapshot to unlock the vm and be deleted in case the
content needs to be deleted first.
Add deleting phase to represent to possible time between the vmsnapshot
marked as to be deleted and the time of waiting for the content to be deleted.

Signed-off-by: Shelly Kagan <skagan@redhat.com>