virt-launcher: add opt-in feature gate for docker bug workaround
A bug in docker (https://github.com/moby/moby/issues/41370) forces
us to implement a workaround in the SELinux options of virt-launcher.
This workaround (disabling categories on non-compute containers)
weakens the security of the project for all when it is only needed
for a very specific use-case: deployments on clusters
that have docker nodes with SELinux enforced on the system and
enabled in the docker configuration (it is disabled by default).

That bug was filed more than 2 years ago, so it is time to disable
the workaround by default, with a feature gate for those who need it.

Signed-off-by: Jed Lejosne <jed@redhat.com>