Migrate index generation process to File-Based Catalog (#2054)
SQLite-Based catlog format, which we were using so far, is being deprecated in favour of the new format of file-based catalog.
With this change, the 'opm index add --bundles' command no longer works, and we need to change the workflow to work with json files inside the catalog image.
References:
- https://olm.operatorframework.io/docs/reference/file-based-catalogs/
- https://github.com/k8s-operatorhub/community-operators/discussions/505

Signed-off-by: Oren Cohen <ocohen@redhat.com>

Signed-off-by: Oren Cohen <ocohen@redhat.com>