instancetype: Create ControllerRevisions earlier within VM reconcile loop
Until now ControllerRevisions for instancetypes and preferences had only
been created by the VirtualMachine controller when starting the
VirtualMachine and initially creating the VirtualMachineInstance. This
delay allowed users to potentially update or delete any referenced
instancetype or preference before a copy was made in a
ControllerRevision.

This change moves the creation of the ControllerRevisions to earlier
within the reconcile loop, ensuring they are created as soon as the
VirtualMachine is seen by the controller. This should help avoid
races with user modifications or removals that could then cause
the VirtualMachine to fail to launch at a later time.

A single VirtualMachine controller unit test is updated as we now expect
the VirtualMachine to be patched with ControllerRevision details ahead
of any conflict check occurring at runtime.

The main ControllerRevision functional test is also updated to assert
that the ControllerRevisions are captured before the VirtualMachine
starts.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>