Mark Reenlightenment VMIs as non-migratable if no TSC frequency exists
VMIs that use HyperV with Reenlightenment (and, implicitly, VMX, as it's
being automatically added to such VMIs) cannot migrate without
explicitly providing TSC frequency to QEMU (more info on that can be
found here: https://github.com/kubevirt/kubevirt/pull/7986).

In Kubevirt we have a mechanism for exposing TSC frequency to the nodes
using node-labeller. However, sometimes (e.g. on some virtual nodes) the
frequency is not exposed. Therefore, VMIs with Reenlightenment shouldn't
be migratable - but they should be able to run.

Signed-off-by: Itamar Holder <iholder@redhat.com>