e2e: fix flaky migration test 1862
There was no special need to use Alpine test tooling image
in the current test. Moreover, the test tooling image requries
some cloud-init configuration in the VMI spec.

I this particular case I am not using libvmi because we
don't have yet builder functions for the DV PVC configuration.

The fix is just return the Alpine image to the simple one, not the tooling.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>