Add restart policy to rsync container used in cluster-sync
An intermittent issue appears when testing kind based e2e jobs with
podman in CI[1]. The rsync container hits an error as the rsync daemon
fails to attach to the specified port. Adding a restart policy to
restart the container any time it stops allows the vgpu lanes to run
successfully using podman.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/logs/periodic-kubevirt-e2e-kind-1.23-vgpu/1564606226711449600

Signed-off-by: Brian Carey <bcarey@redhat.com>