bazel: Update bazeldnf for CentOS Stream 9
The version of bazeldnf we currently use doesn't handle
gracefully the fact that, in CentOS Stream 9, ld-linux-aarch64.so
lives under /usr/lib instead of /usr/lib64. The latter is still
used for ld-linux-x86_64.so even in CentOS Stream 9, and was used
for both in CentOS Stream 8, hence why we hadn't encountered this
limitation until now.

We also need to bring in some packages that contain a caret (^)
in their version number. This is something that the version of
bazeldnf currently in use doesn't handle correctly but the more
recent one does.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>