bazel: Update toolchains for CentOS Stream 9
For x86_64, we simply need to account for the fact that GCC 11
is now used instead of GCC 8.

For aarch64, the cross-building toolchain is not part of the OS
itself but is instead provided as a binary download by Arm. We
pick the most recent release, which is also based on GCC 11; the
version we were using previously was based on GCC 10, so we're
improving consistency across architectures as a side-effect of
this change.

Note that Arm offers the cross-building toolchains under a
slightly different name now, so we also have to account for
that.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>