Introduce PASST_VERSION and update RPMs
As is already the case for other parts of the virtualization
stack like QEMU and libvirt, we want to have very precise control
over the version of passt that is used by KubeVirt. Make it
possible to lock this choice down using an environment variable.

Note that the version of passt is bumped from

  0:0.git.2022_08_04.b516d15-0.el8

to

  0:0.git.2022_08_29.60ffc5b-1.el8

in the process: this was necessary because the former is now
gone from the COPR repo. In order to make the change effective
and keep things consistent, all RPMs are updated.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>