Don't increase the overwrittenModifications metric during upgrade (#2072)
* Don't increase the overwrittenModifications metric during upgrade

Some of the underline custom resources are getting changed as part of the upgrade process.
This causes HCO to increase the overwrittenModifications metric, while this is not a user
action.

This PR changes this behaviour to not increase the metric during upgrade, and so changes made
during upgrade are not count.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* [clean code]: Equal(true/false) => BeTrue()/BeFalse()

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* [clean code]: fix the receiver in methods of OperandHandler and ReconcileHyperConverged

A type method should be either all with pointer receiver or with a regular receiver,
but not both.

This commit set all the methods in the OperandHandler and ReconcileHyperConverged types
to be pointers.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>