Fix hotplug of FileSystem PVCs backed by btrfs
On a host with btrfs the source of the hotplug PVC mount may reside on a
subvolume. E.g. on '/@' like in the example below:

bash-4.4# findmnt -T /local-pvc -N 19272 -J
{
  "filesystems": [
    {
      "target": "/local-pvc",
      "source": "/dev/sda2[/@/opt/local-path-provisioner/pvc-e316b528-46fd-43e8-a64a-ab0e0a54fcca_default_local-pvc]",
      "fstype": "btrfs",
      "options": "rw,relatime,space_cache,subvolid=262,subvol=/@/opt/local-path-provisioner/pvc-e316b528-46fd-43e8-a64a-ab0e0a54fcca_default_local-pvc"
    }
  ]
}

To bind mount such a source, the subvolume '/@' needs to be resolved in
the node's context. This fix reuses the functionality from isolation
package and adds handling of such cases (similar to how it is done with
containerdisks).

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>