Extend bazel sandbox to run commands
Previously container-runtime and build time shared library consistency
was given in `.bazeldnf/sandbox.bazelrc` for `bazel build` and `bazel
test` (since `test` in herits from `build`) commands

This can still lead to situations where `run` commands were building
their tools inside the sandbox but then try to execute these binaries
with the libc.so files from the node.

With the extension of the sandbox mounts to `run` this is not longer the
case. Bazel will always see the same set of shared libraries.

Signed-off-by: Roman Mohr <rmohr@google.com>