instancetype: Support default instance type PVC annotations
This change introduces support for inferring the default instance type
and instance type kind from a specific Volume using the recently
introduced `inferFromVolume` InstancetypeMatcher attribute.

The following combinations are supported by this change:

Volume -> PersistentVolumeClaimVolumeSource -> PersistentVolumeClaim
Volume -> DataVolumeSource -> DataVolumeSourcePVC -> PersistentVolumeClaim
Volume -> DataVolumeSource -> DataVolumeSourceRef -> DataSource -> PersistentVolumeClaim
Volume -> DataVolumeSource -> DataVolumeTemplates -> DataVolumeSourcePVC -> PersistentVolumeClaim
Volume -> DataVolumeSource -> DataVolumeTemplates -> DataVolumeSourceRef -> DataSource -> PersistentVolumeClaim

When a valid Volume name is provided the VirtualMachine mutation webhook
will now attempt to find an underlying PersistentVolumeClaim for the
Volume, checking for the following annotations:

* instancetype.kubevirt.io/default-instancetype

This is required when a Volume name has been provided by
inferFromVolume, failure to find this on the underlying
PersistentVolumeClaim will cause the request to be rejected.

This simply provides the name of the default instance type.

* instancetype.kubevirt.io/default-instancetype-kind

This is optional, when not provided the existing default of
VirtualMachineClusterInstancetype will be applied.

The InstancetypeMatcher of the VirtualMachine is recreated using these
values before the request is passed to the validation webhook where the
existing InstancetypeMatcher checks for conflicts etc are processed.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>