Use the containerdisk image name from the VMI spec
It appears that ContainerStatuses[].Image from the Pod spec cannot be
used to reliably identify the name of the containerdisk image. E.g.
while testing the bellow containerdisk:

    quay.io/kubevirt/fedora-with-test-tooling-container-disk:v0.57.0

it was observed with containerd that ContainerStatuses[].Image may
contain either an internal image ID in the form of:

    sha256:e3559e53b6b1df866598887793f94b9908d3a7e343f4a6394670d69fbe511cb2

or the image can be resolved to an alias:

    192.168.122.1:5000/fedora-with-test-tooling-container-disk@sha256:30750fe1f8936e21890a4e45289ec3afdbcae3d9072ec7d7629ec20f4afbd00e

This eventually breaks live migration, as such references cannot be
resolved on the target node.

To solve the issue, the final pullable reference is now composed using
the base image name from the VMI spec and the unique SHA digest is taken
from ContainerStatuses[].ImageID of the Pod spec.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>