Refactor: Support modifying expected hostname for login functions
This refactoring is introduced in order to change the expected
hostname for login functions.

To make this change generic and basically support all future
modifications / configurations to the login functions,
I've introduced a "LoginConfiguration" interface that is
currently only capable of modifying hostname, but further
configurations can be added in the future which will
prevent the need for another similar refactoring in the
future.

In addition, a default implementation to "LoginConfiguration"
interface is introduced so this change is fully backward
compatible.

Signed-off-by: Itamar Holder <iholder@redhat.com>