network,passt: configure unprivileged_port_start to 0
When using Passt network binding mechanism, and not specifying
any ports explicitly, traffic to the low ports (<1024) is not
passed through to the VM.

This commit sets all virt-launcher ports to be unprivileged, so that
all ports may be used by Passt.

This allows using for example SSH to access VM with Passt binding.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>