expose,e2e: remove tests checking svc after vm removed
Tests removed in this commit create a vm and a service, and
verify that the service doesn't work after the VM is removed.

Testing such behaviour doesn't seem to provide any value to
kubevirt, because it's not KubeVirt responsibility
make a service "not work".

When a VMI/VMIRS/VM is deleted, KubeVirt responsibility is to
delete the virt-launcher pod backing the actual virtual machine.
There is no need to test that k8s Endpoints subset pointing to the
virt-launcher pod(s) is empty, as it's Kubernetes responsibility
to populate k8s Endpoints based on the selector defined in the k8s Service.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>