virt-api: Set default machinetype after preference kind
b91332fd2b1b8b2967b1cac92da0522908e95763 recently added a preference
lookup within the mutation webhook but this was done before
setDefaultPreferenceKind was able to run. As a result the call could
fail if a user had not provided a kind within the preferenceMatcher.

This changes corrects this by moving the setDefaultMachineType to after
setDefaultPreferenceKind ensuring a kind is always set by the mutation
webhook before any lookups are attempted.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>