Fix hook receivers (#2095)
* Fix the hook receivers

According to https://go.dev/doc/faq#methods_on_values_or_pointers, "If
some of the methods of the type must have pointer
receivers, the rest should too, so the method set is consistent
regardless of how the type is used."

Also, remove the receiver name if not used in the method.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

* Remove the reset() mothods from hooks, if not implemented

The reset() method is not part of the hook interface, but it comes from
the reseter interface.

Also, the reset() method is called only if the actual object implements the
reseter interface. So it is safe to remove these methods if there is no
implementation.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>