virt-controller, instancetype: Reintroduce use of informers
The use of these informers was previously removed by
ee4e266dd93b168cc84743a797f64aeb36d75d6b. After further discussions on
the ML [1] however it has become clear that the removal of these
informers from the virt-controller was not required and they can be
reintroduced, this change does this.

The virt-api continues to only use client calls to lookup instancetypes
and preferences.

One small change from the original implementation is that client calls
are now made if an object is not found within an informer. This is to
ensure the object is indeed missing and the failure has not been caused
by a race to update the associated informer.

[1] https://groups.google.com/g/kubevirt-dev/c/q_hR1tFH4Rk

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>