vm-cycle: use exponential backoff when pause due to IO error
Currently, once a VM is Paused due to IO error, the VM status
is flipping between Paused and Running multiple times per second.
This because the virt-handler tries to restart the VM, and if the IOError
is not resolved, the latter will be paused again.
In order to limit the impact to the apiserver, when a domain IOError
occurs, an increasingly exponential backoff will be applied before retrying
to start the VM again.

Signed-off-by: fossedihelm <ffossemo@redhat.com>