Fix bug in restore-controller when creating a restore PVC with both dataSource and dataSourceRef
Since the dataSourceRef API inclusion in k8s v1.22, a new validation mechanism was added for PVCs ensuring that, if present, both the dataSource and dataSourceRef fields must be equal.

This behavior conflicts with the restore-controller, which creates the restore PVC using the source PVC spec as a reference, but overwriting the dataSource field. If the source PVC has a populated dataSourceRef field, just overwriting the dataSource triggers the validation error.

This commit modifies this behavior, so both dataSource and dataSourceRef are overwrritten.

Signed-off-by: Alvaro Romero <alromero@redhat.com>