Refactor: GetPolicyMatchedToVmi() function
GetPolicyMatchedToVmi() had many parameters that made it both
complicated and hard to use. For example, having to provide
a non-nil namespace object even if there is no need in matching
namespace labels. The function is now way simpler with a helper
function dealing with the parallel is a more relaxed way.

Signed-off-by: Itamar Holder <iholder@redhat.com>