e2e,expose: Correct expectation for service endpoints
Endpoint only exposes IP addresses of one family. If dual-stack
cluster is used, then Endpoints expose addresses of whatever family is
the first in the list (for ipv4,ipv6 it's ipv4 and for ipv6,ipv4 it's ipv6).

Since Kubernetes 1.21, there is a new resource EndpointSlices.
The EndpointSlices controller creates a new EndpointSlice for each
address family for Pods selected by a service.

For example, having a Pod with IPv4 and IPv6 address, with a service
selecting it, the following EndpointSlices are created:
$ kubectl get EndpointSlices -n kubevirt-test-default1
NAME                               ADDRESSTYPE   PORTS   ENDPOINTS        AGE
cluster-ip-target-vmivnmz7-l892x   IPv6          80      fd10:244::c494   56m
cluster-ip-target-vmivnmz7-wb6b4   IPv4          80      10.244.196.149   56m

This commit updates e2e test checking the addresses in endpoints so that
the number of addresses in the created EndpointSlices matches the
expected number of IPs (2), instead of asserting the Endpoint to contain
two addresses of two different families.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>