instancetype, virt-api: Use authorizer within expand-spec
This change reworks generateAccessReview of the authorizer into a public
NewSubjectAccessReview function, to make it available for endpoints that
wish to check authorization on CRs handled in their requests.

The functions createInstancetypeResourceAttributes() and
createPreferenceResourceAttributes() are added to the instancetype package.
These functions are used in combination with the refactored authorizer within
the expand-spec endpoint handler to verify the user is allowed to access the
specified instancetype and/or preference before expanding them.

To ensure that every InstanceTypeMatcher and PreferenceMatcher defaults
to the cluster kind if no Kind was specified, the SetDefaultInstancetypeKind
and SetDefaultPreferenceKind methods were moved from the VMsMutator to the
instancetype package and are now used within the non-namespaced expand-spec
endpoint as well.

Co-authored-by: Lee Yarwood <lyarwood@redhat.com>
Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>