snapshot: Restore instancetype controllerrevisions
Building on previous changes that essentially cloned any referenced
instancetype or preference ControllerRevisions during the creation of a
VirtualMachineSnapshot we can now restore these into the existing or new
VirtualMachine.

When restoring for use by a new VirtualMachine we need to ensure that
the newly created ControllerRevisions are claimed by the new
VirtualMachine once it has been created.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>