test 8482: stop expecting migration to finish
It's hard to find the exact timeout / bandwidth limitation configs to
ensure the migration both ends on one hand, and is occuring for enough
time to collect metrics about it. Since the test is not focused on
validating the migration should end, this shouldn't be part of the test
(also, there are many other tests are focused on validating a migration
finish).

In addition, Eventually clause is now changed to keep trying to get
valid metrics of some of them are with a zero-value, something that is
acceptable in Prometheus' context.

Signed-off-by: Itamar Holder <iholder@redhat.com>