Revert "types: Add VirtIO type"
This reverts commit 83aa78efd9b9777fed3a337498f03bd5bafaa407.

A precedent of using device specific typed constants was already present
in the codebase so the use of a single `virtio` constant is pretty
confusing. Instead an additional set of typed constants for the model of
an interface should be introduced with any remaining uses of plain
`virtio` strings also replaced by the existing constant.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>