instancetype: Check contents of existing CRs during restore
This change makes the recently refactored compareRevisions funciton
public and uses it to compare any existing instance type or prefrence
ControllerRevisions found during a VM restore with those associated with
the overall VirtualMachineSnapshot.

If they differ the existing ControllerRevisions are deleted and restored
from the VirtualMachineSnapshot ensuring the VirtualMachine is restored
to the correct state.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>